Crate sc_transaction_pool

Substrate transaction pool implementation.

Modules

error

Transaction pool error.

Structs

BasicPool

Basic implementation of transaction pool that can be customized by providing PoolApi.

FullChainApi

The transaction pool logic for full client.

Options

Pool configuration options.

Pool

Extrinsics pool that performs validation.

PoolLimit

Queue limits

Transaction

Immutable transaction

Enums

RevalidationType

Type of revalidation.

ValidatedTransaction

Pre-validated transaction. Validated pool only accepts transactions wrapped in this enum.

Traits

ChainApi

Concrete extrinsic validation and query logic.

Functions

notification_future

Inform the transaction pool about imported and finalized blocks.

Type Aliases

FullPool

A transaction pool for a full node.