Struct sc_transaction_pool::test_helpers::Pool

pub struct Pool<B: ChainApi> { /* fields omitted */ }

Extrinsics pool that performs validation.

Implementations

impl<B: ChainApi> Pool<B>

pub fn new(options: Options, is_validator: IsValidator, api: Arc<B>) -> Self

Create a new transaction pool.

pub async fn submit_at(
&self,
at: &BlockId<B::Block>,
source: TransactionSource,
xts: impl IntoIterator<Item = ExtrinsicFor<B>>
) -> Result<Vec<Result<<<B as ChainApi>::Block as Block>::Hash, B::Error>>, B::Error>

Imports a bunch of unverified extrinsics to the pool

pub async fn resubmit_at(
&self,
at: &BlockId<B::Block>,
source: TransactionSource,
xts: impl IntoIterator<Item = ExtrinsicFor<B>>
) -> Result<Vec<Result<<<B as ChainApi>::Block as Block>::Hash, B::Error>>, B::Error>

Resubmit the given extrinsics to the pool.

This does not check if a transaction is banned, before we verify it again.

pub async fn submit_one(
&self,
at: &BlockId<B::Block>,
source: TransactionSource,
xt: ExtrinsicFor<B>
) -> Result<<<B as ChainApi>::Block as Block>::Hash, B::Error>

Imports one unverified extrinsic to the pool

pub async fn submit_and_watch(
&self,
at: &BlockId<B::Block>,
source: TransactionSource,
xt: ExtrinsicFor<B>
) -> Result<Watcher<<<B as ChainApi>::Block as Block>::Hash, <<B as ChainApi>::Block as Block>::Hash>, B::Error>

Import a single extrinsic and starts to watch its progress in the pool.

pub fn resubmit(
&self,
revalidated_transactions: HashMap<<<B as ChainApi>::Block as Block>::Hash, ValidatedTransaction<<<B as ChainApi>::Block as Block>::Hash, ExtrinsicFor<B>, <B as ChainApi>::Error>>
)

Resubmit some transaction that were validated elsewhere.

pub fn prune_known(
&self,
at: &BlockId<B::Block>,
hashes: &[<<B as ChainApi>::Block as Block>::Hash]
) -> Result<(), B::Error>

Prunes known ready transactions.

Used to clear the pool from transactions that were part of recently imported block. The main difference from the prune is that we do not revalidate any transactions and ignore unknown passed hashes.

pub async fn prune(
&self,
at: &BlockId<B::Block>,
parent: &BlockId<B::Block>,
extrinsics: &[ExtrinsicFor<B>]
) -> Result<(), B::Error>

Prunes ready transactions.

Used to clear the pool from transactions that were part of recently imported block. To perform pruning we need the tags that each extrinsic provides and to avoid calling into runtime too often we first lookup all extrinsics that are in the pool and get their provided tags from there. Otherwise we query the runtime at the parent block.

pub async fn prune_tags(
&self,
at: &BlockId<B::Block>,
tags: impl IntoIterator<Item = Tag>,
known_imported_hashes: impl IntoIterator<Item = <<B as ChainApi>::Block as Block>::Hash> + Clone
) -> Result<(), B::Error>

Prunes ready transactions that provide given list of tags.

Given tags are assumed to be always provided now, so all transactions in the Future Queue that require that particular tag (and have other requirements satisfied) are promoted to Ready Queue.

Moreover for each provided tag we remove transactions in the pool that:

Provide that tag directly

Are a dependency of pruned transaction.

Returns transactions that have been removed from the pool and must be reverified before reinserting to the pool.

By removing predecessor transactions as well we might actually end up pruning too much, so all removed transactions are reverified against the runtime (validate_transaction) to make sure they are invalid.

However we avoid revalidating transactions that are contained within the second parameter of known_imported_hashes. These transactions (if pruned) are not revalidated and become temporarily banned to prevent importing them in the (near) future.

pub fn hash_of(
&self,
xt: &ExtrinsicFor<B>
) -> <<B as ChainApi>::Block as Block>::Hash

Returns transaction hash

pub fn validated_pool(&self) -> &ValidatedPool<B>

get a reference to the underlying validated pool.