Type Alias sc_transaction_pool::FullPool

pub type FullPool<Block, Client> = BasicPool<FullChainApi<Client, Block>, Block>;

A transaction pool for a full node.

Aliased Type

struct FullPool<Block, Client> { /* private fields */ }

Implementations

impl<Block, Client> FullPool<Block, Client>where Block: BlockT, Client: ProvideRuntimeApi<Block> + BlockBackend<Block> + HeaderBackend<Block> + BlockIdTo<Block> + ExecutorProvider<Block> + UsageProvider<Block> + HeaderMetadata<Block, Error = Error> + Send + Sync + 'static, Client::Api: TaggedTransactionQueue<Block>,

pub fn new_full( options: Options, is_validator: IsValidator, prometheus: Option<&PrometheusRegistry>, spawner: impl SpawnEssentialNamed, client: Arc<Client> ) -> Arc<Self>

Create new basic transaction pool for a full node with the provided api.

impl<PoolApi, Block> BasicPool<PoolApi, Block>where Block: BlockT, PoolApi: ChainApi<Block = Block> + 'static,

pub fn new_test( pool_api: Arc<PoolApi>, best_block_hash: Block::Hash, finalized_hash: Block::Hash ) -> (Self, Pin<Box<dyn Future<Output = ()> + Send>>)

Create new basic transaction pool with provided api, for tests.

pub fn with_revalidation_type( options: Options, is_validator: IsValidator, pool_api: Arc<PoolApi>, prometheus: Option<&PrometheusRegistry>, revalidation_type: RevalidationType, spawner: impl SpawnEssentialNamed, best_block_number: NumberFor<Block>, best_block_hash: Block::Hash, finalized_hash: Block::Hash ) -> Self

Create new basic transaction pool with provided api and custom revalidation type.

pub fn pool(&self) -> &Arc<Pool<PoolApi>>

Gets shared reference to the underlying pool.

pub fn api(&self) -> &PoolApi

Get access to the underlying api