sc_transaction_pool

Type Alias TransactionPoolHandle

pub type TransactionPoolHandle<Block, Client> = TransactionPoolWrapper<Block, Client>;

The public type alias for the actual type providing the implementation of FullClientTransactionPool with the given Client and Block types.

This handle abstracts away the specific type of the transaction pool. Should be used externally to keep reference to transaction pool.

Aliased Type

pub struct TransactionPoolHandle<Block, Client>(pub Box<dyn FullClientTransactionPool<Block, Client, Block = Block, InPoolTransaction = Transaction<<<FullChainApi<Client, Block> as ChainApi>::Block as Block>::Hash, Arc<<<FullChainApi<Client, Block> as ChainApi>::Block as Block>::Extrinsic>>, Hash = <<FullChainApi<Client, Block> as ChainApi>::Block as Block>::Hash, Block = Block, Hash = <<FullChainApi<Client, Block> as ChainApi>::Block as Block>::Hash, Error = <FullChainApi<Client, Block> as ChainApi>::Error, Error = <FullChainApi<Client, Block> as ChainApi>::Error>>);

Tuple Fields
0: Box<dyn FullClientTransactionPool<Block, Client, Block = Block, InPoolTransaction = Transaction<<<FullChainApi<Client, Block> as ChainApi>::Block as Block>::Hash, Arc<<<FullChainApi<Client, Block> as ChainApi>::Block as Block>::Extrinsic>>, Hash = <<FullChainApi<Client, Block> as ChainApi>::Block as Block>::Hash, Block = Block, Hash = <<FullChainApi<Client, Block> as ChainApi>::Block as Block>::Hash, Error = <FullChainApi<Client, Block> as ChainApi>::Error, Error = <FullChainApi<Client, Block> as ChainApi>::Error>>