Crate sc_transaction_pool

Substrate transaction pool implementation.

Re-exports

pub use sc_transaction_graph as txpool;

Modules

error
Transaction pool error.

Structs

BasicPool
Basic implementation of transaction pool that can be customized by providing PoolApi.

FullChainApi
The transaction pool logic for full client.

LightChainApi
The transaction pool logic for light client.

Enums

RevalidationType
Type of revalidation.

Functions

notification_future
Inform the transaction pool about imported and finalized blocks.

Type Definitions

FullPool
A transaction pool for a full node.

LightPool
A transaction pool for a light node.