Crate sc_transaction_pool

Substrate transaction pool implementation.

Macros

insert_and_log_throttled

Inserts a value into a SlidingStats and conditionally logs the current stats using tracing.

insert_and_log_throttled_sync

Sync version of insert_and_log_throttled

Structs

BasicPool

Basic implementation of transaction pool that can be customized by providing PoolApi.

Builder

Builder allowing to create specific instance of transaction pool.

ForkAwareTxPool

The fork-aware transaction pool.

FullChainApi

The transaction pool logic for full client.

Options

Pool configuration options.

Pool

Extrinsics pool that performs validation.

PoolLimit

Queue limits

TimedTransactionSource

A transaction source that includes a timestamp indicating when the transaction was submitted.

TransactionPoolOptions

Transaction pool options.

TransactionPoolWrapper

The wrapper for actual object providing implementation of TransactionPool.

Enums

RevalidationType

Type of revalidation.

TransactionPoolType

The type of transaction pool.

ValidateTransactionPriority

The priority of request to validate the transaction.

Constants

LOG_TARGET

Log target for transaction pool.

Traits

ChainApi

Concrete extrinsic validation and query logic.

Functions

notification_future

Inform the transaction pool about imported and finalized blocks.

Type Aliases

ForkAwareTxPoolTask

Fork aware transaction pool task, that needs to be polled.

TransactionPoolHandle

The public type alias for the actual type providing the implementation of FullClientTransactionPool with the given Client and Block types.