Struct sc_transaction_pool::BasicPool

pub struct BasicPool<PoolApi, Block> where
    Block: BlockT,
    PoolApi: ChainApi<Block = Block>,  { /* fields omitted */ }

Basic implementation of transaction pool that can be customized by providing PoolApi.

Implementations

impl<PoolApi, Block> BasicPool<PoolApi, Block> where
Block: BlockT,
PoolApi: ChainApi<Block = Block> + 'static,

pub fn new_test(
pool_api: Arc<PoolApi>
) -> (Self, Pin<Box<dyn Future<Output = ()> + Send>>, BackSignalControl)

Create new basic transaction pool with provided api, for tests.

pub fn with_revalidation_type(
options: Options,
is_validator: IsValidator,
pool_api: Arc<PoolApi>,
prometheus: Option<&PrometheusRegistry>,
revalidation_type: RevalidationType,
spawner: impl SpawnEssentialNamed,
best_block_number: NumberFor<Block>
) -> Self

Create new basic transaction pool with provided api and custom revalidation type.

pub fn pool(&self) -> &Arc<Pool<PoolApi>>

Gets shared reference to the underlying pool.

pub fn api(&self) -> &PoolApi

Get access to the underlying api

impl<Block, Client, Fetcher> BasicPool<LightChainApi<Client, Fetcher, Block>, Block> where
Block: BlockT,
Client: HeaderBackend<Block> + UsageProvider<Block> + 'static,
Fetcher: Fetcher<Block> + 'static,

pub fn new_light(
options: Options,
prometheus: Option<&PrometheusRegistry>,
spawner: impl SpawnEssentialNamed,
client: Arc<Client>,
fetcher: Arc<Fetcher>
) -> Self

Create new basic transaction pool for a light node with the provided api.

impl<Block, Client> BasicPool<FullChainApi<Client, Block>, Block> where
Block: BlockT,
Client: ProvideRuntimeApi<Block> + BlockBackend<Block> + HeaderBackend<Block> + BlockIdTo<Block> + ExecutorProvider<Block> + UsageProvider<Block> + Send + Sync + 'static,
Client::Api: TaggedTransactionQueue<Block>,

pub fn new_full(
options: Options,
is_validator: IsValidator,
prometheus: Option<&PrometheusRegistry>,
spawner: impl SpawnEssentialNamed,
client: Arc<Client>
) -> Arc<Self>

Create new basic transaction pool for a full node with the provided api.